added card class
